Do not require OpenMM for tests
