+++ openff/toolkit/_tests/test_molecule.py
@@ -24,6 +24,7 @@ import numpy as np
 import pytest
 from openff.units import unit
 from openff.units.elements import MASSES, SYMBOLS
+from openff.utilities import skip_if_missing
 
 from openff.toolkit._tests.create_molecules import (
     create_acetaldehyde,
@@ -3918,6 +3919,7 @@ class TestMoleculeResiduePerception:
             assert found == expected
 
 
+@skip_if_missing("openmm")
 class TestMoleculeFromPDB:
     """
     Test creation of cheminformatics-rich openff Molecule from PDB files.
@@ -4559,6 +4561,7 @@ class TestHierarchies:
                 iterator_name="atoms",
             )
 
+    @skip_if_missing("openmm")
     def test_add_default_hierarchy_schemes(self):
         """Test add_default_hierarchy_schemes and its kwargs"""
         offmol = Molecule.from_polymer_pdb(
