+++ setup.cfg
@@ -32,6 +32,7 @@ per-file-ignores =
     openff/toolkit/_tests/test_molecule.py
     openff/toolkit/_tests/test_forcefield.py:E501
     openff/toolkit/_tests/test_toolkits.py:E501
+    openff/toolkit/_tests/utils.py:F401
 
 [isort]
 profile=black
